requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
canvas-lms-api @ git+https://github.com/Mizzou-CS-Core/CanvasRequestLibrary.git
    # via -r requirements.in
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via -r requirements.in
smmap==5.0.2
    # via gitdb
tomlkit==0.13.2
    # via -r requirements.in
